create middleware isAdmin and protected routes

+++ routes/web.php
@@ -1,9 +1,10 @@
 <?php
 
+use App\Models\Product;
+use App\Http\Middleware\IsAdmin;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\ProductController;
-use App\Models\Product;
 
 
 
@@ -30,12 +31,12 @@ Auth::routes();
 Route::get('/', [ProductController::class, 'index'])->name('home');
 
 Route::get('/show/{id}', [ProductController::class, 'show'])->name('show');
-Route::post('/products', [ProductController::class, 'store'])->name('store');
-Route::delete('/delete/{id}', [ProductController::class, 'destroy'])->name('delete');
-Route::patch('/products/{id}', [ProductController::class, 'update'])->name('update');
-Route::get('/edit/{id}', [ProductController::class, 'edit'])->name('edit');
+Route::post('/products', [ProductController::class, 'store'])->name('store')->middleware(IsAdmin::class);
+Route::delete('/delete/{id}', [ProductController::class, 'destroy'])->name('delete')->middleware(IsAdmin::class);
+Route::patch('/products/{id}', [ProductController::class, 'update'])->name('update')->middleware(IsAdmin::class);
+Route::get('/edit/{id}', [ProductController::class, 'edit'])->name('edit')->middleware(IsAdmin::class);
 
-Route::get('/create', [ProductController::class, 'create'])->name('create');
+Route::get('/create', [ProductController::class, 'create'])->name('create')->middleware(IsAdmin::class);
 
 Route::get('/search', [ProductController::class, 'search'])->name('search');
